Benchmarking/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0
scanpy>=1.8.0
anndata>=0.7.0
PyYAML>=5.4.0
